rabbitmq/rabbitmq_managment/requirements.txt
loguru==0.7.2
pika==1.3.2
PyYAML==6.0.1
requests==2.31.0
